+++ packages/master/src/templates/master/master.module.ts
@@ -1,4 +1,3 @@
-import { CommandModule } from '@mbc-cqrs-serverless/core'
 import { MasterModule as CoreMasterModule } from '@mbc-cqrs-serverless/master'
 import { Module } from '@nestjs/common'
 import { PrismaService } from 'src/prisma'
@@ -7,10 +6,6 @@ import { MasterDataSyncRdsHandler } from './handler/master-rds.handler'
 
 @Module({
   imports: [
-    CommandModule.register({
-      tableName: 'master',
-      dataSyncHandlers: [MasterDataSyncRdsHandler],
-    }),
     CoreMasterModule.register({
       enableController: true,
       prismaService: PrismaService,